requirements.txt
Flask==2.0.1
Pillow==8.4.0
python-dotenv==0.19.0
Werkzeug==2.0.1
gunicorn==20.1.0 